+++ src/ParticleContainer.cpp
@@ -12,8 +12,9 @@ void ParticleContainer::addParticle(const Particle* p) {
   particles.emplace_back(*p);
 }
 
-void ParticleContainer::removeParticle(const int idx) {
-  if (idx < particles.size()) particles.erase(particles.begin() + idx);
+void ParticleContainer::removeParticle(const size_t idx) {
+  if (idx < particles.size())
+    particles.erase(particles.begin() + idx);
 }
 
 ParticleContainer::iterator ParticleContainer::begin() {
